2.7.0 - The Game Status Update
* now only suggests channels that you can read and send to when you try
to run a command in the wrong channel.
+ GroupUp command
+ Groups Channel setting added
+ ClearGroupUp command
+ SetPlayingStatus command
* fixed a bug involving the generation of the Guild Files.
* started work on wiki page generation.
+ playing statuses that randomly select a game from chucklefish's
published games, starbound 2/5, Wargroove 1/5, all others 1/25
+ added website blacklist check to url variables.
+ QuickRespond DM command
+ started implementation of Sail's Pixel xp system, (is now tracking XP)
* Fixed a bug preventing messages from being deleted when adding the X
emoji as a reaction to a message.
+ added a system for users to add up to 5 links to their profile via the
EditLinks command.
+ XpDenied Channel Setting adde

+++ src/main/java/Handlers/FileHandler.java
@@ -64,15 +64,22 @@ public class FileHandler {
     /**
      * Writes to File using "file" as the path.
      */
-    public static void writeToFile(String file, String text) {
+    public static void writeToFile(String file, String text,boolean overwrite) {
         try {
             if (!Files.exists(Paths.get(file))) {
                 Files.createFile(Paths.get(file));
             }
-            FileWriter fileWriter = new FileWriter(file, true);
-            fileWriter.append(text + "\n");
-            fileWriter.flush();
-            fileWriter.close();
+            if (overwrite){
+                FileWriter fileWriter = new FileWriter(file, false);
+                fileWriter.write(text);
+                fileWriter.flush();
+                fileWriter.close();
+            }else {
+                FileWriter fileWriter = new FileWriter(file, true);
+                fileWriter.append(text + "\n");
+                fileWriter.flush();
+                fileWriter.close();
+            }
             logger.trace("Writing to file: " + file);
         } catch (IOException e) {
             logger.error(e.getCause().toString());
@@ -119,6 +126,6 @@ public class FileHandler {
     }
 
     public static void initFile(String path) {
-        if (!exists(path)) writeToFile(path, "");
+        if (!exists(path)) writeToFile(path, "",false);
     }
 }
